Add GNU compiler options to COMPILE_OPTIONS for each target. #17
A separate list of command line options is defined for both xcompact3d
and the decomp2d library. Because they are currently the same, the
list could technically be defined one for decomp2d with the PUBLIC
attribute. However, it is likely that these 2 lists will diverge in
the future. Or at least it's helpful to allow it.